requirements.txt
azure-storage-blob
azure-identity
requests
pyyaml